feat: enhance task and workflow management with real-time updates and default workflow seeding

+++ backend/Velyo.Application/Workflows/Queries/GetWorkflowStates/WorkflowStateDto.cs
@@ -4,4 +4,4 @@ public record WorkflowStateDto(
     Guid Id,
     string Name,
     string Color,
-    float OrderIndex);
+    int OrderIndex);
